Move FirstStepFormConcerns to application controller
These methods are not a concern. They are methods shared by the two
controllers which manage the two possible first steps.
